lambda/requirements.txt
google-api-python-client==2.188.0
google-auth==2.48.0
pygithub==2.8.1